# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Restaurant QR Table Ordering System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tables:
    get:
      operationId: getTables
      tags: [tables]
      summary: List all tables
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of tables
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Table"
    post:
      operationId: createTable
      tags: [tables]
      summary: Create a new table
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTableRequest"
      responses:
        "201":
          description: Table created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Table"
  /tables/{tableId}:
    get:
      operationId: getTable
      tags: [tables]
      summary: Get table details
      security:
        - bearerAuth: []
      parameters:
        - name: tableId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Table info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Table"
  /tables/{tableId}/qr:
    get:
      operationId: getTableQr
      tags: [tables]
      summary: Get QR code URL for table
      security:
        - bearerAuth: []
      parameters:
        - name: tableId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: QR code data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QrResponse"
  /session/validate:
    get:
      operationId: validateSession
      tags: [tables]
      summary: Validate guest session token from QR
      parameters:
        - name: token
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Session valid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionInfo"
        "401":
          description: Invalid or expired session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /reports/summary:
    get:
      operationId: getReportSummary
      tags: [reports]
      summary: Get sales summary report
      security:
        - bearerAuth: []
      parameters:
        - name: from
          in: query
          schema:
            type: string
            format: date
        - name: to
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Sales summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportSummary"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
...
    Table:
      type: object
      properties:
        id:
          type: integer
        number:
          type: integer
        name:
          type: string
        capacity:
          type: integer
        status:
          type: string
          enum: [available, occupied, reserved]
        qrToken:
          type: string
      required: [id, number, name, capacity, status]
    CreateTableRequest:
      type: object
      properties:
        number:
          type: integer
        name:
          type: string
        capacity:
          type: integer
      required: [number, name, capacity]
    QrResponse:
      type: object
      properties:
        tableId:
          type: integer
        tableName:
          type: string
        token:
          type: string
        url:
          type: string
      required: [tableId, tableName, token, url]
    SessionInfo:
      type: object
      properties:
        tableId:
          type: integer
        tableName:
          type: string
        tableNumber:
          type: integer
        valid:
          type: boolean
      required: [tableId, tableName, tableNumber, valid]
...
    ReportSummary:
      type: object
      properties:
        totalOrders:
          type: integer
        totalRevenue:
          type: number
        averageOrderValue:
          type: number
        ordersByStatus:
          type: object
          additionalProperties:
            type: integer
        topItems:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              quantity:
                type: integer
              revenue:
                type: number
      required: [totalOrders, totalRevenue, averageOrderValue]